requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
alembic
pydantic-settings
pytest
pytest-asyncio
httpx
greenlet